refactor: move library extension to shared interface
