+++ tools/serverpod_cli/lib/src/generator/code_generator.dart
@@ -1,4 +1,7 @@
+import 'package:code_builder/code_builder.dart';
+import 'package:dart_style/dart_style.dart';
 import 'package:serverpod_cli/analyzer.dart';
+import 'package:serverpod_cli/src/logger/logger.dart';
 
 /// A code generator is responsible for generating the code for the target
 /// language.
@@ -33,4 +36,25 @@ abstract class CodeGenerator {
     required ProtocolDefinition protocolDefinition,
     required GeneratorConfig config,
   });
+}
+
+extension GenerateCode on Library {
+  String generateCode() {
+    var code = accept(DartEmitter.scoped(useNullSafetySyntax: true)).toString();
+    try {
+      return DartFormatter().format('''
+/* AUTOMATICALLY GENERATED CODE DO NOT MODIFY */
+/*   To generate run: "serverpod generate"    */
+
+// ignore_for_file: library_private_types_in_public_api
+// ignore_for_file: public_member_api_docs
+// ignore_for_file: implementation_imports
+
+$code
+''');
+    } on FormatterException catch (e) {
+      log.error(e.toString());
+    }
+    return code;
+  }
 }
